fix: убрать FeedController, метод для получения ленты перенести в UserController
